feat: implement favorites feature with models, repository, and view components
