+++ lib/features/favorites/repository/favorites_remote_data_source.dart
@@ -1,11 +1,10 @@
 import 'package:sponti/core/constants/api_constants.dart';
 import 'package:sponti/core/errors/exceptions.dart';
-import 'package:sponti/features/locations/model/location_model.dart';
+import 'package:sponti/features/favorites/model/favorite_model.dart';
 import 'package:supabase_flutter/supabase_flutter.dart' hide AuthException;
 
 abstract interface class FavoritesRemoteDataSource {
-  Future<List<String>> getFavoriteLocationIds();
-  Future<List<LocationModel>> getFavoriteLocations();
+  Future<List<FavoriteModel>> getFavorites();
   Future<void> addFavorite(String locationId);
   Future<void> removeFavorite(String locationId);
 }
@@ -24,41 +23,16 @@ class FavoritesRemoteDataSourceImpl implements FavoritesRemoteDataSource {
   }
 
   @override
-  Future<List<String>> getFavoriteLocationIds() async {
+  Future<List<FavoriteModel>> getFavorites() async {
     try {
       final response = await _client
           .from(ApiConstants.favoritesTable)
-          .select('location_id')
+          .select('location_id, user_id, created_at, locations(*)')
           .eq('user_id', _userId)
           .order('created_at', ascending: false);
 
       return (response as List<dynamic>)
-          .map((row) => (row as Map<String, dynamic>)['location_id'])
-          .whereType<String>()
-          .toList(growable: false);
-    } on AuthException {
-      rethrow;
-    } on PostgrestException catch (e) {
-      throw ServerException(e.message);
-    } catch (e) {
-      throw ServerException(e.toString());
-    }
-  }
-
-  @override
-  Future<List<LocationModel>> getFavoriteLocations() async {
-    try {
-      final response = await _client
-          .from(ApiConstants.favoritesTable)
-          .select('created_at, locations(*)')
-          .eq('user_id', _userId)
-          .order('created_at', ascending: false);
-
-      return (response as List<dynamic>)
-          .map((row) => row as Map<String, dynamic>)
-          .map((row) => row['locations'])
-          .whereType<Map<String, dynamic>>()
-          .map(LocationModel.fromJson)
+          .map((row) => FavoriteModel.fromJson(row as Map<String, dynamic>))
           .toList(growable: false);
     } on AuthException {
       rethrow;
